Updated resources files
