--- src/components/Resources/Resources.js
@@ -1,171 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import classNames from 'classnames';
-import AppBar from '@material-ui/core/AppBar';
-import Button from '@material-ui/core/Button';
-import CameraIcon from '@material-ui/icons/PhotoCamera';
-import Card from '@material-ui/core/Card';
-import CardActions from '@material-ui/core/CardActions';
-import CardContent from '@material-ui/core/CardContent';
-import CardMedia from '@material-ui/core/CardMedia';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import Grid from '@material-ui/core/Grid';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-import { withStyles } from '@material-ui/core/styles';
-import CustomizedInputBase from '../CustomizedInputBase'
-import Search from "../Search";
-import axios from "axios";
-
-class Resources extends Component {
-  
-    constructor(props) {
-      super(props);
-      this.state = {
-        open: false,
-        restaurants: []
-      };
-    }
-    handleClickOpen = () => {
-      this.setState({ open: true });
-    };
-
-    handleClose = () => {
-      this.setState({ open: false });
-    };
-
-    componentDidMount() {
-      axios.get("/restaurants")
-          .then(res => {
-              const restaurant = res.data;
-              this.setState({ restaurant });
-          })
-    };
-
-    handleInputChange = event => {
-      // Getting the value and name of the input which triggered the change
-      
-      let value = event.target.value;
-      const name = event.target.name;
-      console.log(event.target);
-      // Updating the input's state
-      this.setState({
-        [name]: value
-      });
-    };
-  
-    handleSubmit = event => {
-      event.preventDefault();
-
-      const restaurant = {
-          website: this.state.website,
-          thumbnail: this.state.thumbnail,
-          name: this.state.name,
-          rating: this.state.rating,
-          summary: this.state.summary,
-      };
-
-      axios.post("/post/restaurants",  restaurant )
-          .then(res => {
-              console.log(res);
-              console.log(res.data);
-          })
-          .catch((err) => {
-              console.log(err);
-          })
-  };
-
-
-    render() {
-    const { classes } = this.props;
-    return (
-      
-    <React.Fragment>
-      <CssBaseline />
-      <AppBar position="static" className={classes.appBar}>
-        <Toolbar>
-          <CameraIcon className={classes.icon} />
-          <Typography variant="h6" color="inherit" noWrap>
-            Dog-Friendly Places to Go
-          </Typography>
-        </Toolbar>
-      </AppBar>
-      <main>
-        {/* Hero unit */}
-        <div className={classes.heroUnit}>
-          <div className={classes.heroContent}>
-            <Typography component="h1" variant="h2" align="center" color="textPrimary" gutterBottom>
-              Album layout
-            </Typography>
-            <Typography variant="h6" align="center" color="textSecondary" paragraph>
-              Resources: Dog-friendly restaurants, hotels and parks            </Typography>
-            <div className={classes.heroButtons}>
-              <Grid container spacing={16} justify="center">
-                <Grid item>
-                  <Button variant="contained" color="primary">
-                  //<CustomizedInputBase/> //get location to search
-                  // <Search/> //Retrieve the server data based on Location from CustomizedInputbase
-
-                  </Button>
-                </Grid>
-                <Grid item>
-                  <Button variant="outlined" color="primary">
-                    Secondary action
-                  </Button>
-                </Grid>
-              </Grid>
-            </div>
-          </div>
-        </div>
-        <div className={classNames(classes.layout, classes.cardGrid)}>
-          {/* End hero unit */}
-          <Grid container spacing={40}>
-            {cards.map(card => (
-              <Grid item key={card} sm={6} md={4} lg={3}>
-                <Card className={classes.card}>
-                  <CardMedia
-                    className={classes.cardMedia}
-                    image="data:image/svg+xml;charset=UTF-8,%3Csvg%20width%3D%22288%22%20height%3D%22225%22%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20viewBox%3D%220%200%20288%20225%22%20preserveAspectRatio%3D%22none%22%3E%3Cdefs%3E%3Cstyle%20type%3D%22text%2Fcss%22%3E%23holder_164edaf95ee%20text%20%7B%20fill%3A%23eceeef%3Bfont-weight%3Abold%3Bfont-family%3AArial%2C%20Helvetica%2C%20Open%20Sans%2C%20sans-serif%2C%20monospace%3Bfont-size%3A14pt%20%7D%20%3C%2Fstyle%3E%3C%2Fdefs%3E%3Cg%20id%3D%22holder_164edaf95ee%22%3E%3Crect%20width%3D%22288%22%20height%3D%22225%22%20fill%3D%22%2355595c%22%3E%3C%2Frect%3E%3Cg%3E%3Ctext%20x%3D%2296.32500076293945%22%20y%3D%22118.8%22%3EThumbnail%3C%2Ftext%3E%3C%2Fg%3E%3C%2Fg%3E%3C%2Fsvg%3E" // eslint-disable-line max-len
-                    title="Image title"
-                  />
-                  <CardContent className={classes.cardContent}>
-                    <Typography gutterBottom variant="h5" component="h2">
-                      Restaurant
-                    </Typography>
-                    <Typography>
-                      Dog-friendly restaurants near you
-                    </Typography>
-                  </CardContent>
-                  <CardActions>
-                    <Button size="small" color="primary">
-                      View
-                    </Button>
-                    <Button size="small" color="primary">
-                      Edit
-                    </Button>
-                  </CardActions>
-                </Card>
-              </Grid>
-            ))}
-          </Grid>
-        </div>
-      </main>
-      {/* Footer */}
-      <footer className={classes.footer}>
-        <Typography variant="h6" align="center" gutterBottom>
-          Footer
-        </Typography>
-        <Typography variant="subtitle1" align="center" color="textSecondary" component="p">
-          Created by an amazing dog-lover team!
-        </Typography>
-      </footer>
-      {/* End footer */}
-    </React.Fragment>
-  );
-}
-
-Resources.propTypes = {
-  classes: PropTypes.object.isRequired,
-};
-
-export default withStyles(styles)(Resources);
