--- src/components/Resources/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Resources.js";
